complete code

+++ lib/UI/ProfileTabPage/ProfileTabPage.dart
@@ -1,23 +1,28 @@
 
+import 'package:com.zat.linedup/Components/ScaffoldMessageWidget/ScaffoldMessageWidget.dart';
+import 'package:com.zat.linedup/Controllers/LogoutService/LogoutService.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:linedup_app/Components/CacheNetworkImage/CacheNetworkImage.dart';
-import 'package:linedup_app/Components/CustomAppButton/CustomAppButton.dart';
-import 'package:linedup_app/Components/Extentions/PaddingExtentions.dart';
-import 'package:linedup_app/Components/ProfilleBackgroundImage/ProfileBackgroundeImage.dart';
-import 'package:linedup_app/Components/StaticTextStyle/StaticTextStyle.dart';
-import 'package:linedup_app/Controllers/UserProfileService/UserProfileService.dart';
-import 'package:linedup_app/Providers/UserProfileProvider/UserProfileProvider.dart';
-import 'package:linedup_app/SharedPrefrences/SharedPrefrences.dart';
-import 'package:linedup_app/Utils/Constants/AssetConstants/AssetConstants.dart';
-import 'package:linedup_app/Utils/Constants/ColorConstants/ColorConstants.dart';
-import 'package:linedup_app/Utils/Constants/Key_Constants.dart';
-import 'package:linedup_app/Utils/Constants/RouteConstants/RouteConstants.dart';
-import 'package:linedup_app/globals.dart';
+import 'package:com.zat.linedup/Components/CacheNetworkImage/CacheNetworkImage.dart';
+import 'package:com.zat.linedup/Components/CustomAppButton/CustomAppButton.dart';
+import 'package:com.zat.linedup/Components/Extentions/PaddingExtentions.dart';
+import 'package:com.zat.linedup/Components/ProfilleBackgroundImage/ProfileBackgroundeImage.dart';
+import 'package:com.zat.linedup/Components/StaticTextStyle/StaticTextStyle.dart';
+import 'package:com.zat.linedup/Controllers/UserProfileService/UserProfileService.dart';
+import 'package:com.zat.linedup/Providers/SocialLoginsAuthProvider/SocailLoginsAuthProvider.dart';
+import 'package:com.zat.linedup/Providers/UserProfileProvider/UserProfileProvider.dart';
+import 'package:com.zat.linedup/SharedPrefrences/SharedPrefrences.dart';
+import 'package:com.zat.linedup/Utils/Constants/AssetConstants/AssetConstants.dart';
+import 'package:com.zat.linedup/Utils/Constants/ColorConstants/ColorConstants.dart';
+import 'package:com.zat.linedup/Utils/Constants/Key_Constants.dart';
+import 'package:com.zat.linedup/Utils/Constants/RouteConstants/RouteConstants.dart';
+import 'package:com.zat.linedup/globals.dart';
 import 'package:provider/provider.dart';
+
+import '../../Components/CustomWaveClipper.dart';
 
 class ProfileTabPage extends StatefulWidget {
   const ProfileTabPage({super.key});
@@ -32,6 +37,7 @@ class _ProfileTabPageState extends State<ProfileTabPage> {
   void initState() {
     // TODO: implement initState
     super.initState();
+
     userProfileService.getUserProfile(context);
   }
   @override
@@ -39,8 +45,8 @@ class _ProfileTabPageState extends State<ProfileTabPage> {
 
     return Scaffold(
 
-      body: Consumer<UserProfileProvider>(
-        builder: (context, userProfileProvider,_) {
+      body: Consumer2<UserProfileProvider,SocialAuthProvider>(
+        builder: (context, userProfileProvider,socialAuthProvider,_) {
           var userProfile = userProfileProvider.userProfileResponseModel;
           return userProfile?.data==null?Center(
             child: CircularProgressIndicator(
@@ -48,46 +54,62 @@ class _ProfileTabPageState extends State<ProfileTabPage> {
             ),
           ) : Column(
             children: [
-              SizedBox(
-                child: Stack(
-                  alignment: AlignmentDirectional.bottomCenter,
-                  children: [
-                    Container(
-                      height: 240.sp,
+              Stack(
+                children: [
+                  ClipPath(
+                    // clipper: WaveClipper(),
+                    child: Container(
+                      height: 227.sp,
                       width: double.infinity,
-                      child: Image.asset(AssetConstants.profileBackGroundImage,fit: BoxFit.fitHeight,height: 220.sp,),
+                      color: ColorConstants.appPrimaryColor,
+                      // child: Image.asset(AssetConstants.profileBackGroundImage,fit: BoxFit.fitHeight,height: 220.sp,),
                     ),
+                  ),
 
-                    Container(
+                  Padding(
+                    padding:  EdgeInsets.only(top: 90.sp,left: 100.sp),
+                    child: Container(
                       height: 155.sp,
                       width: 155.sp,
                       decoration: BoxDecoration(
-                        border: Border.all(color: ColorConstants.appPrimaryColor,width: 1.sp),
-                        shape: BoxShape.circle
+                          border: Border.all(color: ColorConstants.appPrimaryColor,width: 1.sp),
+                          shape: BoxShape.circle
                       ),
                       child: ClipOval(
                         clipBehavior: Clip.antiAlias,
-                        child: ImageWidget(
+                        child: socialAuthProvider.currentUser?.photoURL!=null?ImageWidget(
                           height: 150.sp,
                           width: 150.sp,
                           fit: BoxFit.cover,
                           decoration: BoxDecoration(
-                            shape: BoxShape.circle
+                              shape: BoxShape.circle
+                          ),
+                          imageUrl: socialAuthProvider.currentUser?.photoURL,
+
+                        ) :ImageWidget(
+                          height: 150.sp,
+                          width: 150.sp,
+                          fit: BoxFit.cover,
+                          decoration: BoxDecoration(
+                              shape: BoxShape.circle
                           ),
                           imageUrl: userProfile?.data?.profileImage.toString(),
                         ),
                       ),
                     ),
-                  ],
-                ),
+                  ),
+                ],
               ),
-
-              Expanded(
+              Container(
                 child: Padding(
-                  padding:  EdgeInsets.only(top: 30.sp,right: 20.sp,left:20.sp),
+                  padding:  EdgeInsets.only(top: 20.sp,right: 20.sp,left:20.sp),
                   child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.center,
                     children: [
-                      Text(userProfile?.data?.userName??"",style: StaticTextStyle().boldTextStyle.copyWith(
+                     socialAuthProvider.currentUser?.displayName!=null?Text( socialAuthProvider.currentUser?.displayName??"",style: StaticTextStyle().boldTextStyle.copyWith(
+                         fontSize: 18.sp,
+                         color: ColorConstants.profileNameColor
+                     ),)  :Text(userProfile?.data?.userName??"",style: StaticTextStyle().boldTextStyle.copyWith(
                         fontSize: 18.sp,
                         color: ColorConstants.profileNameColor
                       ),),
@@ -95,7 +117,7 @@ class _ProfileTabPageState extends State<ProfileTabPage> {
                         fontSize: 13.sp,
                         color: ColorConstants.textGreyColor
                       ),),
-                      Text("Gender, ${userProfile?.data?.gender??""}",style: StaticTextStyle().regular.copyWith(
+                      Text("Gender ${userProfile?.data?.gender??""}",style: StaticTextStyle().regular.copyWith(
                         fontSize: 13.sp,
                         color: ColorConstants.textGreyColor
                       ),),
@@ -183,8 +205,21 @@ class _ProfileTabPageState extends State<ProfileTabPage> {
                         padding:  EdgeInsets.symmetric(vertical: 30.sp),
                         child: InkWell(
                           onTap: ()async{
-                            await SharedPreferencesService().remove(KeysConstants.accessToken, token);
-                            Navigator.of(context).pushNamed(RouteConstants.loginPageRoute);
+                            if(userProfile?.data?.googleId!=null){
+                              await SharedPreferencesService().remove(KeysConstants.accessToken, token);
+                              socialAuthProvider.handleSignOut();
+                              Navigator.of(context).pushNamedAndRemoveUntil(RouteConstants.loginPageRoute,arguments: "logout", (route) => false);
+
+                            }else{
+                              var response = await LogoutService().logoutUser(context);
+                              if(response!.responseData!.success==true){
+                                ShowMessage().showMessage(context, "Logout!", ColorConstants.redColor);
+                                Navigator.of(context).pushNamedAndRemoveUntil(RouteConstants.loginPageRoute,arguments: "logout", (route) => false);
+
+                              }
+
+                            }
+
                           },
                           child: Row(
                             children: [
